Change TO dses to separate versions
Puts all the logic in deliveryservices.go, and makes v12.go, v13.go,
v14.go purely boilerplate.

This should make it significantly easier to add new versions: you'll
only have to copy the previous v13.go file, rename v13->v14, and
add your logic to the main deliveryservices.go

Also adds a section on adding new versions to the README.md.
